fix(examples): build the dashboard on FreeBSD, OpenBSD and NetBSD
Statfs_t field types differ per platform — Bavail is int64 on FreeBSD — and
OpenBSD and NetBSD have no Blocks field or no Statfs at all, yet the unix build
tag sent all of them to the same file. Convert both operands explicitly and
limit that file to linux, darwin and freebsd; the others take the fallback.

+++ examples/dashboard/disk_unix.go
@@ -1,4 +1,4 @@
-//go:build unix || darwin || linux
+//go:build linux || darwin || freebsd
 
 package main
 
@@ -11,8 +11,9 @@ func getDiskSpace() (usedGB, totalGB, percent float64) {
 		return 0, 0, 0
 	}
 
-	totalBytes := stat.Blocks * uint64(stat.Bsize)
-	freeBytes := stat.Bavail * uint64(stat.Bsize)
+	totalBytes := uint64(stat.Blocks) * uint64(stat.Bsize)
+	// The field types differ between platforms (Bavail is int64 on FreeBSD).
+	freeBytes := uint64(stat.Bavail) * uint64(stat.Bsize)
 	if totalBytes == 0 {
 		return 0, 0, 0
 	}
